pyproject.toml
[project]
name =  "livesrt"
version = "0.1.0"
description = "Live subtitles for your life"
authors = [
    { name = "Rémy Sanchez", email = "remy.sanchez@hyperthese.net" },
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "rich",
    "rich-click",
    "click",
    "httpx",
    "pyyaml",
    "keyring",
    "pwinput",
    "pyaudio",
    "websockets",
    "pydantic",
    "tenacity",
    "huggingface-hub~=1.2.1",
    "llama-cpp-python~=0.3.16",
    "textual>=6.8.0",
    "dependency-injector>=4.48.3",
]

[project.scripts]
livesrt = "livesrt.cli:cli"

[dependency-groups]
dev = [
    "mypy",
    "ruff",
    "pytest",
    "types-pyaudio",
    "ipython",
    "pytest-asyncio>=1.3.0",
    "pytest-httpx>=0.36.0",
]

[build-system]
requires = ["uv_build>=0.9.5,<0.10.0"]
build-backend = "uv_build"

[tool.ruff.lint]
select = [
    # -- Core --
    "E",  # pycodestyle (errors)
    "W",  # pycodestyle (warnings)
    "F",  # Pyflakes
    "I",  # isort
    "C90",  # McCabe Complexity
    "D1",  # Mandatory docstrings
    "S",  # Bandit

    # -- Bonus --
    "A",  # flake8-bugbear
    "B",  # flake8-bugbear
    "DTZ",  # date/time zone
    "T10",  # debugger
    "EM",  # error messages
    "EXE",  # executable
    "G",  # logging
    "UP",  # pyupgrade
    "ASYNC",  # async
    "INP",  # PEP 420
    "T20",  # no print
    "PT",  # pytest
    "SLOT",  # slots
    "TCH",  # type checking blocks
    "PTH",  # Use pathlib
    "FIX",  # Don't allow not done stuff in code
    "ERA",  # Don't comment code
    "RUF",  # Ruff stuff
    "TID251",  # See below
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D", "UP", "S101"]

[tool.ruff.lint.flake8-tidy-imports.banned-api]
"requests".msg = "Use `httpx` instead, which has timeouts by default."

[tool.ruff.lint.pydocstyle]
convention = "numpy"

[tool.mypy]
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "pwinput"
ignore_missing_imports = true
